Moved the test files to the same location as their respective source files, in order to keep the same standard of "fetch-location-handler.test.ts"
